add i2c module
For now, only I2C target with an EEPROM like emulation is supported. The
I2C slave mode is somewhat awkward with this controller as it only
really supports slave mode with clock stretching. OTOH, we really want
to avoid clock strechting as it might break the I2C bus and put it into
a stuck state where even the master can't release it from. For this, we
need to supply all the (transmitting) data ahead of time and an EEPROM
like interface is the easiest method for this.

Signed-off-by: Michael Walle <michael.walle@kontron.com>