add SMARC watchdog module
The MCU is responsible to drive the WDT_TIME_OUT# pin. Add a watchdog
module to do that. To ease the linux and uboot driver development,
emulate the watchdog module of the SMARC-sAL28 module. To simplify it
even more, we use a hardware timer and the software just steers it. This
will also make it less prone to software errors because most
functionality runs on the bare hardware.

The original watchdog runs with a frequency of 1Hz. The hardware timers
on the MSP0L can't run at this low speed. Fortunately, the original
watchdog had an 8bit range and the hardware timers have 16bit counters.
This way, we can misuse the lower 8bit as a third prescaler. The upper
bits are then interpreted as the actual watchdog counter value.

Signed-off-by: Michael Walle <michael.walle@kontron.com>